Add a Transformer for the ScottyWindowOperator in Kafka Streams.
The ScottyWindowOperator implements the Proccessor interface.
Therefore, it is only possible to call the .process(...) function from the
Kafka Streams DSL, which is a terminating function.
Thus, it is not possible to do further computations.
With the new class, which implement the transformer interface, it is possible
to call .transform(...) on the stream and do further computations.