Add Supplier for the Processor and Transformer and use them in the kafka streams demo.
From the docs:
The supplier should always generate a new instance each time get() gets called.
Creating a single Transformer object and returning the same object reference in get()
would be a violation of the supplier pattern and leads to runtime exceptions.

Executing the demo with more than one testInput topic, a runtime exception is thrown.
Using a supplier fixes this issue.