Fix bug that same key is returned for different window results
The processWatermark(..) function is called inside the process(..) function
and used the key of the process function to write back to Kafka.
However, windows with different can be processed there and produce a result.
This may result in the same key being used for different window results.
Using the key of the given SlicingWindowOperator leads to the desired result.